+++ docs/changes.md
@@ -48,6 +48,7 @@ Changes in command line arguments:
 - [NXDRIVE-1225](https://jira.nuxeo.com/browse/NXDRIVE-1225): Be specific about pyobjc requirements
 - [NXDRIVE-1234](https://jira.nuxeo.com/browse/NXDRIVE-1234): Tweak pyup.io parameters
 - [NXDRIVE-1244](https://jira.nuxeo.com/browse/NXDRIVE-1244): Ensure Drive is closed before auto-upgrading on Windows
+- [NXDRIVE-1323](https://jira.nuxeo.com/browse/NXDRIVE-1323): Use venv instead of virtualenv on Windows
 
 ### Doc
 - [NXDRIVE-1276](https://jira.nuxeo.com/browse/NXDRIVE-1276): Add documentation about Nuxeo Platform support